refactor: 수익률 계산 로직 Service 레이어로 이동
